Update LICENSE
I may modify a couple things about the MIT license in the future, but I don't think it matters for this repository, so I'll leave as-is for now. I realized having creative commons is weird here. I was used to it because it's the license I use on my music.